Fix window resizing error
