+++ forui/lib/src/foundation/rendering.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/rendering.dart';
+import 'package:meta/meta.dart';
+
+@internal
+extension RenderBoxes on RenderBox {
+  BoxParentData get data => parentData! as BoxParentData;
+}
