+++ forui/lib/src/widgets/slider/inherited_data.dart
@@ -17,6 +17,7 @@ final class InheritedData extends InheritedWidget {
   final FSliderStyle style;
   final Layout layout;
   final List<FSliderMark> marks;
+  final double? trackMainAxisExtent;
   final double? trackHitRegionCrossExtent;
   final Widget Function(FTooltipStyle, double) tooltipBuilder;
   final String Function(FSliderSelection) semanticFormatterCallback;
@@ -27,6 +28,7 @@ final class InheritedData extends InheritedWidget {
     required this.style,
     required this.layout,
     required this.marks,
+    required this.trackMainAxisExtent,
     required this.trackHitRegionCrossExtent,
     required this.tooltipBuilder,
     required this.semanticFormatterCallback,
@@ -41,6 +43,7 @@ final class InheritedData extends InheritedWidget {
       style != old.style ||
       layout != old.layout ||
       !marks.equals(marks) ||
+      trackMainAxisExtent != old.trackMainAxisExtent ||
       trackHitRegionCrossExtent != old.trackHitRegionCrossExtent ||
       tooltipBuilder != old.tooltipBuilder ||
       semanticFormatterCallback != old.semanticFormatterCallback ||
@@ -54,6 +57,7 @@ final class InheritedData extends InheritedWidget {
       ..add(DiagnosticsProperty('style', style))
       ..add(EnumProperty('layout', layout))
       ..add(IterableProperty('marks', marks))
+      ..add(DoubleProperty('trackMainAxisExtent', trackMainAxisExtent))
       ..add(DoubleProperty('trackHitRegionCrossExtent', trackHitRegionCrossExtent))
       ..add(ObjectFlagProperty.has('tooltipBuilder', tooltipBuilder))
       ..add(ObjectFlagProperty.has('semanticFormatterCallback', semanticFormatterCallback))
