+++ forui/lib/src/widgets/slider/slider.dart
@@ -1,9 +1,11 @@
+import 'package:collection/collection.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 
 import 'package:forui/forui.dart';
+import 'package:forui/src/widgets/slider/form_field.dart';
+import 'package:forui/src/widgets/slider/inherited_controller.dart';
 import 'package:forui/src/widgets/slider/inherited_data.dart';
-import 'package:forui/src/widgets/slider/layout.dart';
 
 /// An input where the user selects a value from within a given range.
 ///
@@ -16,7 +18,7 @@ import 'package:forui/src/widgets/slider/layout.dart';
 /// * [FDiscreteSliderController.new] for selecting a discrete value.
 /// * [FDiscreteSliderController.range] for selecting a discrete range.
 /// * [FSliderStyles] for customizing a slider's appearance.
-class FSlider extends FormField<FSliderSelection> {
+class FSlider extends StatelessWidget {
   static Widget _errorBuilder(BuildContext context, String error) => Text(error);
 
   static Widget _tooltipBuilder(FTooltipStyle _, double value) => Text('${(value * 100).toStringAsFixed(0)}%');
@@ -78,6 +80,44 @@ class FSlider extends FormField<FSliderSelection> {
   // ignore: avoid_positional_boolean_parameters
   final String Function(double) semanticValueFormatterCallback;
 
+  /// An optional method to call with the final value when the form is saved via [FormState.save].
+  final FormFieldSetter<String>? onSaved;
+
+  /// An optional method that validates an input. Returns an error string to
+  /// display if the input is invalid, or null otherwise.
+  ///
+  /// The returned value is exposed by the [FormFieldState.errorText] property. It transforms the text using
+  /// [errorBuilder].
+  ///
+  /// Alternating between error and normal state can cause the height of the [FTextField] to change if no other
+  /// subtext decoration is set on the field. To create a field whose height is fixed regardless of whether or not an
+  /// error is displayed, wrap the [FTextField] in a fixed height parent like [SizedBox].
+  final FormFieldValidator<String>? validator;
+
+  /// Used to enable/disable this form field auto validation and update its error text.
+  ///
+  /// Defaults to [AutovalidateMode.disabled].
+  ///
+  /// If [AutovalidateMode.onUserInteraction], this FormField will only auto-validate after its content changes. If
+  /// [AutovalidateMode.always], it will auto-validate even without user interaction. If [AutovalidateMode.disabled],
+  /// auto-validation will be disabled.
+  final AutovalidateMode? autovalidateMode;
+
+  /// An optional property that forces the [FormFieldState] into an error state
+  /// by directly setting the [FormFieldState.errorText] property without
+  /// running the validator function.
+  ///
+  /// When the [forceErrorText] property is provided, the [FormFieldState.errorText]
+  /// will be set to the provided value, causing the form field to be considered
+  /// invalid and to display the error message specified.
+  ///
+  /// When [validator] is provided, [forceErrorText] will override any error that it
+  /// returns. [validator] will not be called unless [forceErrorText] is null.
+  final String? forceErrorText;
+
+  /// True if the slider is enabled.
+  final bool enabled;
+
   /// Creates a [FSlider].
   FSlider({
     required this.controller,
@@ -92,80 +132,43 @@ class FSlider extends FormField<FSliderSelection> {
     this.tooltipBuilder = _tooltipBuilder,
     this.semanticValueFormatterCallback = _semanticValueFormatter,
     String Function(FSliderSelection)? semanticFormatterCallback,
-    super.onSaved,
-    super.validator,
-    super.forceErrorText,
-    super.enabled = true,
-    super.autovalidateMode,
-    super.restorationId,
+    this.onSaved,
+    this.validator,
+    this.forceErrorText,
+    this.enabled = true,
+    this.autovalidateMode,
     super.key,
-  })  : semanticFormatterCallback = semanticFormatterCallback ?? _formatter(controller),
-        super(
-          builder: (field) {
-            final state = field as _State;
-            final styles = state.context.theme.sliderStyles;
-            final sliderStyle = style ?? (layout.vertical ? styles.verticalStyle : styles.horizontalStyle);
-            final (labelState, stateStyle) = switch ((state.hasError, enabled)) {
-              (false, true) => (FLabelState.enabled, sliderStyle.enabledStyle),
-              (false, false) => (FLabelState.disabled, sliderStyle.disabledStyle),
-              (true, _) => (FLabelState.error, sliderStyle.errorStyle),
-            };
-
-            return InheritedData(
-              style: sliderStyle,
-              stateStyle: stateStyle,
-              layout: layout,
-              marks: marks,
-              trackHitRegionCrossExtent: trackHitRegionCrossExtent,
-              enabled: enabled,
-              tooltipBuilder: tooltipBuilder,
-              semanticFormatterCallback: semanticFormatterCallback ?? _formatter(controller),
-              semanticValueFormatterCallback: semanticValueFormatterCallback,
-              child: LayoutBuilder(
-                builder: (context, constraints) => SliderLayout(
-                  controller: controller,
-                  layoutStyle: sliderStyle.labelLayoutStyle,
-                  stateStyle: stateStyle,
-                  state: labelState,
-                  layout: layout,
-                  label: label == null
-                      ? const SizedBox()
-                      : DefaultTextStyle(
-                          style: stateStyle.labelTextStyle,
-                          child: Padding(
-                            padding: sliderStyle.labelLayoutStyle.labelPadding,
-                            child: label,
-                          ),
-                        ),
-                  description: description == null
-                      ? const SizedBox()
-                      : DefaultTextStyle.merge(
-                          style: stateStyle.descriptionTextStyle,
-                          child: Padding(
-                            padding: sliderStyle.labelLayoutStyle.descriptionPadding,
-                            child: description,
-                          ),
-                        ),
-                  error: state.errorText == null
-                      ? const SizedBox()
-                      : DefaultTextStyle.merge(
-                    style: (stateStyle as FSliderErrorStyle).errorTextStyle,
-                        child: Padding(
-                            padding: sliderStyle.labelLayoutStyle.errorPadding,
-                            child: errorBuilder(context, state.errorText!),
-                          ),
-                      ),
-                  marks: marks,
-                  constraints: constraints,
-                  mainAxisExtent: trackMainAxisExtent,
-                ),
-              ),
-            );
-          },
-        );
+  }) : semanticFormatterCallback = semanticFormatterCallback ?? _formatter(controller);
 
   @override
-  FormFieldState<FSliderSelection> createState() => _State();
+  Widget build(BuildContext context) {
+    final styles = context.theme.sliderStyles;
+    final sliderStyle = style ?? (layout.vertical ? styles.verticalStyle : styles.horizontalStyle);
+    return InheritedData(
+      style: sliderStyle,
+      layout: layout,
+      marks: marks,
+      trackMainAxisExtent: trackMainAxisExtent,
+      trackHitRegionCrossExtent: trackHitRegionCrossExtent,
+      enabled: enabled,
+      tooltipBuilder: tooltipBuilder,
+      semanticFormatterCallback: semanticFormatterCallback,
+      semanticValueFormatterCallback: semanticValueFormatterCallback,
+      child: LayoutBuilder(
+        builder: (context, constraints) => _Slider(
+          controller: controller,
+          style: sliderStyle,
+          layout: layout,
+          label: label,
+          description: description,
+          errorBuilder: errorBuilder,
+          marks: marks,
+          constraints: constraints,
+          mainAxisExtent: trackMainAxisExtent,
+        ),
+      ),
+    );
+  }
 
   @override
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
@@ -174,64 +177,113 @@ class FSlider extends FormField<FSliderSelection> {
       ..add(DiagnosticsProperty('controller', controller))
       ..add(DiagnosticsProperty('style', style))
       ..add(EnumProperty('layout', layout))
+      ..add(ObjectFlagProperty.has('errorBuilder', errorBuilder))
       ..add(IterableProperty('marks', marks))
       ..add(DoubleProperty('trackMainAxisExtent', trackMainAxisExtent))
       ..add(DoubleProperty('trackHitRegionCrossExtent', trackHitRegionCrossExtent))
-      ..add(FlagProperty('enabled', value: enabled, ifTrue: 'enabled', ifFalse: 'disabled'))
       ..add(ObjectFlagProperty.has('tooltipBuilder', tooltipBuilder))
       ..add(ObjectFlagProperty.has('semanticFormatterCallback', semanticFormatterCallback))
-      ..add(ObjectFlagProperty.has('semanticValueFormatterCallback', semanticValueFormatterCallback));
+      ..add(ObjectFlagProperty.has('semanticValueFormatterCallback', semanticValueFormatterCallback))
+      ..add(ObjectFlagProperty.has('onSaved', onSaved))
+      ..add(ObjectFlagProperty.has('validator', validator))
+      ..add(FlagProperty('enabled', value: enabled, ifTrue: 'enabled', ifFalse: 'disabled'))
+      ..add(EnumProperty('autovalidateMode', autovalidateMode))
+      ..add(ObjectFlagProperty.has('forceErrorText', forceErrorText));
   }
 }
 
-class _State extends FormFieldState<FSliderSelection> {
+class _Slider extends StatefulWidget {
+  final FSliderController controller;
+  final FSliderStyle style;
+  final Layout layout;
+  final Widget? label;
+  final Widget? description;
+  final Widget Function(BuildContext, String) errorBuilder;
+  final List<FSliderMark> marks;
+  final BoxConstraints constraints;
+  final double? mainAxisExtent;
+
+  const _Slider({
+    required this.controller,
+    required this.style,
+    required this.layout,
+    required this.label,
+    required this.description,
+    required this.errorBuilder,
+    required this.marks,
+    required this.constraints,
+    required this.mainAxisExtent,
+  });
+
+  @override
+  State<_Slider> createState() => _SliderState();
+
+  double get _mainAxisExtent {
+    final insets = style.labelLayoutStyle.childPadding;
+    final extent = switch (mainAxisExtent) {
+      final extent? => extent,
+      _ when layout.vertical => constraints.maxHeight - insets.top - insets.bottom,
+      _ => constraints.maxWidth - insets.left - insets.right,
+    };
+
+    if (extent.isInfinite) {
+      throw FlutterError(
+        switch (layout.vertical) {
+          true => 'A vertical FSlider was given an infinite height although it needs a finite height. To fix this, '
+              'consider supplying a mainAxisExtent or placing FSlider in a SizedBox.',
+          false => 'A horizontal FSlider was given an infinite width although it needs a finite width. To fix this, '
+              'consider supplying a mainAxisExtent or placing FSlider in a SizedBox.',
+        },
+      );
+    }
+
+    return extent;
+  }
+
+  @override
+  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
+    super.debugFillProperties(properties);
+    properties
+      ..add(DiagnosticsProperty('controller', controller))
+      ..add(DiagnosticsProperty('style', style))
+      ..add(EnumProperty('layout', layout))
+      ..add(ObjectFlagProperty.has('errorBuilder', errorBuilder))
+      ..add(IterableProperty('marks', marks))
+      ..add(DiagnosticsProperty('constraints', constraints))
+      ..add(DoubleProperty('mainAxisExtent', mainAxisExtent));
+  }
+}
+
+class _SliderState extends State<_Slider> {
   @override
   void initState() {
     super.initState();
-    widget.controller.addListener(_handleControllerChanged);
+    widget.controller.attach(widget._mainAxisExtent, widget.marks);
   }
 
   @override
-  void didUpdateWidget(covariant FSlider old) {
+  void didUpdateWidget(covariant _Slider old) {
     super.didUpdateWidget(old);
-    if (widget.controller == old.controller) {
-      return;
-    }
-
-    widget.controller.addListener(_handleControllerChanged);
-    old.controller.removeListener(_handleControllerChanged);
-  }
-
-  @override
-  void didChange(FSliderSelection? value) {
-    // TODO: fix bug when resizing window.
-    // This is not 100% accurate since a controller's selection can never be null. However, users will have to go out
-    // of their way to obtain a FormFieldState<FSliderSelection> via a GlobalKey AND call didChange(null).
-    assert(value != null, "A slider's selection cannot be null.");
-    super.didChange(value);
-    if (widget.controller.selection != value) {
-      widget.controller.selection = value;
+    if (widget.controller != old.controller ||
+        widget.layout != old.layout ||
+        widget._mainAxisExtent != old._mainAxisExtent ||
+        !widget.marks.equals(old.marks)) {
+      widget.controller.attach(widget._mainAxisExtent, widget.marks);
     }
   }
 
   @override
-  void reset() {
-    // Set the controller value before calling super.reset() to let _handleControllerChanged suppress the change.
-    widget.controller.reset();
-    super.reset();
-  }
-
-  void _handleControllerChanged() {
-    // Suppress changes that originated from within this class.
-    //
-    // In the case where a controller has been passed in to this widget, we register this change listener. In these
-    // cases, we'll also receive change notifications for changes originating from within this class -- for example, the
-    // reset() method. In such cases, the FormField value will already have been set.
-    if (widget.controller.selection != value) {
-      didChange(widget.controller.selection);
-    }
-  }
-
-  @override
-  FSlider get widget => super.widget as FSlider;
+  Widget build(BuildContext context) => ListenableBuilder(
+    listenable: widget.controller,
+    builder: (context, _) => InheritedController(
+      controller: widget.controller,
+      child: SliderFormField(
+        controller: widget.controller,
+        constraints: widget.constraints,
+        label: widget.label,
+        description: widget.description,
+        errorBuilder: widget.errorBuilder,
+      ),
+    ),
+  );
 }
